requirements.txt
# 项目核心依赖
discord.py[speed]
python-dotenv
Pillow
requests
aiohttp
aiosqlite
google-genai
google-api-core
PyYAML
chromadb
objgraph
uvloop; sys_platform != 'win32'

# Web UI 依赖
fastapi
uvicorn
gunicorn
eventlet
docker
psutil

# 测试依赖
pytest
pytest-asyncio
pytest-dotenv